+++ sources/OpenMcdf/StreamRW.cs
@@ -41,7 +41,7 @@ namespace OpenMcdf
         public int ReadInt32()
         {
             this.stream.Read(buffer, 0, 4);
-            return (int)(buffer[0] | (buffer[1] << 8) | (buffer[2] << 16) | (buffer[3] << 24));
+            return buffer[0] | (buffer[1] << 8) | (buffer[2] << 16) | (buffer[3] << 24);
         }
 
         public uint ReadUInt32()
